requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
python-multipart
anthropic
chromadb
sentence-transformers
langchain
langchain-text-splitters
langchain-core
pypdf
python-docx
streamlit
plotly
pandas
python-dotenv
pydantic-settings
email-validator
tqdm
tenacity

# Testing dependencies
pytest==7.4.3
pytest-html==4.1.1
